feat: configured Spring security
* feat: created migrations to work with Spring security

* feat: added dependencies and configured Spring Security

* feat: added AuthAdapter to get authorized user

* refactor: optimized resume dto's

* refactor: improved create, edit and delete logic with using AuthAdapter

* refactor: optimized vacancy dto's

* refactor: improved create, edit and delete logic with using AuthAdapter

* chore: moved AuthAdapter to utils

* feat: added new role admin and data

* refactor: improved logic to get resume by Id

* feat: added sql validation exception handler

* feat: added classes to check role's existence

* refactor: changed everything related to user due to changes in db

* refactor: improved edit and delete logic with using AuthAdapter

* feat: handled Multipart and IllegalState errors

* feat: added classes for responding to vacancies

* feat: handled new category and response exceptions

* all done

---------

+++ src/main/java/kg/attractor/job_search/service/impl/UserServiceImpl.java
@@ -9,6 +9,7 @@ import kg.attractor.job_search.exception.EmployerNotFoundException;
 import kg.attractor.job_search.exception.UserNotFoundException;
 import kg.attractor.job_search.mapper.UserMapper;
 import kg.attractor.job_search.model.User;
+import kg.attractor.job_search.service.RoleService;
 import kg.attractor.job_search.service.UserService;
 import kg.attractor.job_search.util.FileUtil;
 import lombok.RequiredArgsConstructor;
@@ -16,6 +17,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 import org.springframework.web.multipart.MultipartFile;
@@ -27,6 +30,8 @@ import java.util.List;
 @RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
     private final UserDao userDao;
+    private final RoleService roleService;
+    private final PasswordEncoder encoder;
 
     @Override
     public List<UserDto> getUsers() {
@@ -81,13 +86,16 @@ public class UserServiceImpl implements UserService {
             throw new IllegalArgumentException("Пользователь с таким email уже существует");
         }
 
+        roleService.getRoleId(userDto.getRoleId());
+
         String userName = userDto.getName().trim().toLowerCase();
         userName = StringUtils.capitalize(userName);
 
         userDto.setName(userName);
         userDto.setEmail(userDto.getEmail().trim().toLowerCase());
         userDto.setPhoneNumber(userDto.getPhoneNumber().trim().toLowerCase());
-        userDto.setAccountType(userDto.getAccountType().trim().toLowerCase());
+        userDto.setRoleId(userDto.getRoleId());
+        userDto.setPassword(encoder.encode(userDto.getPassword()));
 
         Long id = userDao.registerUser(UserMapper.toUser(userDto)) ;
         log.info("Register user: {}", id);
@@ -97,6 +105,10 @@ public class UserServiceImpl implements UserService {
     @Override
     public Long updateUser(Long userId, EditUserDto userDto) {
         getUserById(userId);
+
+        if(!userId.equals(userDto.getId())) {
+            throw new AccessDeniedException("Вы не имеете права на редактироание чужого профиля!");
+        }
 
         String name = userDto.getName().trim().toLowerCase();
         name = StringUtils.capitalize(name);
@@ -114,8 +126,11 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public HttpStatus deleteUser(Long userId) {
+    public HttpStatus deleteUser(Long userId, Long authId) {
         getUserById(userId);
+        if(!userId.equals(authId)) {
+            throw new AccessDeniedException("Вы не имеете права удалять чужжой профиль!");
+        }
         userDao.deleteUser(userId);
         log.info("Deleted user: {}", userId);
         return HttpStatus.OK;
@@ -156,7 +171,10 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public MultipartFile uploadAvatar(Long userId, MultipartFile file) {
+    public MultipartFile uploadAvatar(Long userId, MultipartFile file, Long authId) {
+        if(!userId.equals(authId)) {
+            throw new AccessDeniedException("Вы не имеете права на загрузку аватара другому профилю!");
+        }
         String contentType = file.getContentType();
         if (contentType == null || (!contentType.equals("image/jpeg") && !contentType.equals("image/png"))) {
             throw new IllegalArgumentException("Только файлы JPEG и PNG разрешены для загрузки");
